// frontend/generated/com/flightmonitor/application/endpoint/entity/Area.ts
/**
 * This module is generated from com.flightmonitor.application.endpoint.entity.Area.
 * All changes to this file are overridden. Please consider to make changes in the corresponding Java file if necessary.
 * @see {@link file:///home/ladnyik/vaadin/flight-monitor/src/main/java/com/flightmonitor/application/endpoint/entity/Area.java}
 */

export default interface Area {
  bottomLatitude: number;
  bottomLongitude: number;
  topLatitude: number;
  topLongitude: number;
}